# backend/src/resources/feature/dto/update-feature.dto.ts
import { IsBoolean, IsOptional, IsString } from 'class-validator';

export class UpdateFeatureDto {
  @IsString()
  @IsOptional()
  label?: string;

  @IsBoolean()
  @IsOptional()
  isEnabled?: boolean;
}