# backend/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"Successfully fetched hello message"}},"summary":"Get a hello message","tags":["app"]}},"/protected":{"get":{"operationId":"AppController_getProtectedResource","parameters":[],"responses":{"200":{"description":"Successfully fetched protected resource"},"401":{"description":"Unauthorized access"}},"summary":"Get a protected resource","tags":["app"]}},"/users":{"get":{"operationId":"UserController_findAll","parameters":[{"name":"search","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved all users"}},"summary":"Retrieve all users","tags":["users"]},"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"The user has been successfully created."},"400":{"description":"Invalid input data"}},"summary":"Create a new user","tags":["users"]}},"/users/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved user"},"404":{"description":"User not found"}},"summary":"Retrieve a single user by ID","tags":["users"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"User successfully deleted"},"404":{"description":"User not found"}},"summary":"Delete a user by ID","tags":["users"]}},"/users/{id}/teams":{"put":{"operationId":"UserController_updateUserTeams","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Teams successfully updated"},"404":{"description":"User not found"}},"summary":"Update user teams","tags":["users"]}},"/users/{id}/team/{teamId}":{"put":{"operationId":"UserController_setCurrentTeam","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"teamId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Current team successfully updated"},"404":{"description":"User or team not found"}},"summary":"Set user current team","tags":["users"]}},"/users/{userId}/teams/{teamId}":{"post":{"operationId":"UserController_addTeamToUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"teamId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["users"]}},"/teams":{"get":{"operationId":"TeamController_findAll","parameters":[],"responses":{"200":{"description":"Successfully retrieved all teams"}},"summary":"Retrieve all teams","tags":["teams"]},"post":{"operationId":"TeamController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTeamDto"}}}},"responses":{"201":{"description":"The team has been successfully created."},"400":{"description":"Invalid input data"}},"summary":"Create a new team","tags":["teams"]}},"/teams/{id}":{"get":{"operationId":"TeamController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved team"},"404":{"description":"Team not found"}},"summary":"Retrieve a single team by ID","tags":["teams"]},"patch":{"operationId":"TeamController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTeamDto"}}}},"responses":{"200":{"description":"Team successfully updated"},"404":{"description":"Team not found"}},"summary":"Update a team by ID","tags":["teams"]},"delete":{"operationId":"TeamController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Team successfully deleted"},"404":{"description":"Team not found"}},"summary":"Delete a team by ID","tags":["teams"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"User successfully logged in"},"401":{"description":"Invalid credentials"}},"summary":"Login user","tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"Returns the authenticated user"}},"summary":"Get authenticated user information","tags":["auth"]}},"/features":{"post":{"operationId":"FeatureController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFeatureDto"}}}},"responses":{"201":{"description":"The feature has been successfully created."},"400":{"description":"Invalid input data"}},"summary":"Create a new feature","tags":["feature"]},"get":{"operationId":"FeatureController_findAll","parameters":[{"name":"search","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved all features"}},"summary":"Retrieve all features","tags":["feature"]}},"/features/{id}":{"get":{"operationId":"FeatureController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved feature"},"404":{"description":"Feature not found"}},"summary":"Retrieve a feature by ID","tags":["feature"]},"patch":{"operationId":"FeatureController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFeatureDto"}}}},"responses":{"200":{"description":"Feature has been updated successfully"},"404":{"description":"Feature not found"}},"summary":"Update a feature by ID","tags":["feature"]},"delete":{"operationId":"FeatureController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Feature successfully removed"},"404":{"description":"Feature not found"}},"summary":"Remove a feature by ID","tags":["feature"]}},"/infos":{"post":{"operationId":"InfoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInfoDto"}}}},"responses":{"201":{"description":"The info has been successfully created."},"400":{"description":"Invalid input data"}},"summary":"Create new info","tags":["infos"]},"get":{"operationId":"InfoController_findAll","parameters":[],"responses":{"200":{"description":"Successfully retrieved all infos"}},"summary":"Retrieve all infos","tags":["infos"]}},"/infos/{id}":{"get":{"operationId":"InfoController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved info"},"404":{"description":"Info not found"}},"summary":"Retrieve a single info by ID","tags":["infos"]},"patch":{"operationId":"InfoController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInfoDto"}}}},"responses":{"200":{"description":"Info has been updated successfully"},"404":{"description":"Info not found"}},"summary":"Update info by ID","tags":["infos"]},"delete":{"operationId":"InfoController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Info successfully removed"},"404":{"description":"Info not found"}},"summary":"Remove info by ID","tags":["infos"]}}},"info":{"title":"Tuile","description":"Documentation de l'API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"CreateTeamDto":{"type":"object","properties":{}},"UpdateTeamDto":{"type":"object","properties":{}},"CreateFeatureDto":{"type":"object","properties":{}},"UpdateFeatureDto":{"type":"object","properties":{}},"CreateInfoDto":{"type":"object","properties":{}},"UpdateInfoDto":{"type":"object","properties":{}}}}}